requirements.txt
httpx
tqdm
termcolor
requests
python-nmap
dnspython
PyYAML
